refactor(docs): snippet dx
