+++ apps/docs/content/components/snippet/colors.raw.jsx
@@ -1,0 +1,14 @@
+import {Snippet} from "@nextui-org/react";
+
+export default function App() {
+  return (
+    <div className="flex flex-wrap gap-4">
+      <Snippet color="default">npm install @nextui-org/react</Snippet>
+      <Snippet color="primary">npm install @nextui-org/react</Snippet>
+      <Snippet color="secondary">npm install @nextui-org/react</Snippet>
+      <Snippet color="success">npm install @nextui-org/react</Snippet>
+      <Snippet color="warning">npm install @nextui-org/react</Snippet>
+      <Snippet color="danger">npm install @nextui-org/react</Snippet>
+    </div>
+  );
+}
