+++ apps/docs/content/components/snippet/custom-tooltip.raw.jsx
@@ -1,0 +1,17 @@
+import {Snippet} from "@nextui-org/react";
+
+export default function App() {
+  return (
+    <Snippet
+      tooltipProps={{
+        color: "foreground",
+        content: "Copy this snippet",
+        disableAnimation: true,
+        placement: "right",
+        closeDelay: 0,
+      }}
+    >
+      npm install @nextui-org/react
+    </Snippet>
+  );
+}
