+++ apps/docs/content/components/snippet/custom-tooltip.ts
@@ -1,20 +1,4 @@
-const App = `import {Snippet} from "@nextui-org/react";
-
-export default function App() {
-  return (
-    <Snippet
-      tooltipProps={{
-        color: "foreground",
-        content: "Copy this snippet",
-        disableAnimation: true,
-        placement: "right",
-        closeDelay: 0
-      }}
-    >
-      npm install @nextui-org/react
-    </Snippet>
-  );
-}`;
+import App from "./custom-tooltip.raw.jsx?raw";
 
 const react = {
   "/App.jsx": App,
