+++ apps/docs/content/components/snippet/multiline.ts
@@ -1,14 +1,4 @@
-const App = `import {Snippet} from "@nextui-org/react";
-
-export default function App() {
-  return (
-    <Snippet>
-      <span>npm install @nextui-org/react</span>
-      <span>yarn add @nextui-org/react</span>
-      <span>pnpm add @nextui-org/react</span>
-    </Snippet>
-  );
-}`;
+import App from "./multiline.raw.jsx?raw";
 
 const react = {
   "/App.jsx": App,
