+++ apps/docs/content/components/snippet/sizes.ts
@@ -1,14 +1,4 @@
-const App = `import {Snippet} from "@nextui-org/react";
-
-export default function App() {
-  return (
-    <div className="flex flex-wrap gap-4">
-      <Snippet size="sm">npm install @nextui-org/react</Snippet>
-      <Snippet size="md">npm install @nextui-org/react</Snippet>
-      <Snippet size="lg">npm install @nextui-org/react</Snippet>
-    </div>  
-  );
-}`;
+import App from "./sizes.raw.jsx?raw";
 
 const react = {
   "/App.jsx": App,
