+++ apps/docs/content/components/snippet/symbol.raw.jsx
@@ -1,0 +1,9 @@
+import {Snippet} from "@nextui-org/react";
+
+export default function App() {
+  return (
+    <Snippet symbol="#" variant="bordered">
+      npm install @nextui-org/react
+    </Snippet>
+  );
+}
