+++ apps/docs/content/components/snippet/usage.raw.jsx
@@ -1,0 +1,5 @@
+import {Snippet} from "@nextui-org/react";
+
+export default function App() {
+  return <Snippet>npm install @nextui-org/react</Snippet>;
+}
