+++ apps/docs/content/components/snippet/variants.ts
@@ -1,15 +1,4 @@
-const App = `import {Snippet} from "@nextui-org/react";
-
-export default function App() {
-  return (
-    <div className="flex flex-wrap gap-4">
-      <Snippet variant="bordered">npm install @nextui-org/react</Snippet>
-      <Snippet variant="flat" color="warning">npm install @nextui-org/react</Snippet>
-      <Snippet variant="solid" color="primary">npm install @nextui-org/react</Snippet>
-      <Snippet variant="shadow" color="secondary">npm install @nextui-org/react</Snippet>
-    </div> 
-  );
-}`;
+import App from "./variants.raw.jsx?raw";
 
 const react = {
   "/App.jsx": App,
