+++ apps/docs/content/components/snippet/without-copy.raw.jsx
@@ -1,0 +1,9 @@
+import {Snippet} from "@nextui-org/react";
+
+export default function App() {
+  return (
+    <Snippet hideCopyButton color="primary" variant="solid">
+      npm install @nextui-org/react
+    </Snippet>
+  );
+}
